backend/requirements.txt
fastapi
uvicorn
pandas
numpy
fairlearn
scikit-learn
google-genai
python-multipart
pydantic
python-dotenv
reportlab
